refactor: Jpa관련 config 수정
<description>
<issue-number>
